requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
openai==1.55.3
pypdf==3.17.1
streamlit==1.41.1
tiktoken==0.5.2
